refactor: 정적 파일 핸들링 서블릿 패키지 변경
